Packets fixes WIndows

+++ common/src/packet/PacketCreateRoom.cpp
@@ -17,7 +17,7 @@ PacketCreateRoom::PacketCreateRoom(const std::string & gameName, const uint8_t &
 
 	ps.add((uint16_t)_gameName.size());
 	ps.add(_gameName);
-	dataPacketSize += 2 + _gameName.size();
+	dataPacketSize += 2 + (uint32_t)_gameName.size();
 
 	ps.add(_maxPlayers);
 	dataPacketSize += 1;
@@ -40,7 +40,7 @@ PacketCreateRoom::PacketCreateRoom(const uint8_t *data)
 		_data[a] = data[a + 9];
 
 	_gameName = pd.getString(posInPacket + 2, pd.get16(posInPacket));
-	posInPacket += 2 + pd.get16(posInPacket);
+	posInPacket += 2 + (uint32_t)pd.get16(posInPacket);
 
 	_maxPlayers = pd.get8(posInPacket);
 	posInPacket += 1;
